+++ conreality_headset/CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.2.1 - 2019-07-30
+### Changed
+- Added a dependency on AndroidX core classes
+- Declared the Android permissions the plugin needs
+### Added
+- Support for recording audio from Bluetooth SCO headsets (on Android only)
+
 ## 0.2.0 - 2019-07-15
 ### Changed
 - Required Flutter 1.7.8+ and Dart 2.4.0+